+++ src/XRoadLib/System/Web/Services/Description/SoapBinding.cs
@@ -1,17 +1,25 @@
 #if NETSTANDARD1_5
 
 using System.Xml;
+using XRoadLib;
 
 namespace System.Web.Services.Description
 {
     public class SoapBinding : ServiceDescriptionFormatExtension
     {
-        public SoapBindingStyle Style { get; set; }
-        public string Transport { get; set; }
+        public SoapBindingStyle Style { get; set; } = SoapBindingStyle.Document;
+        public string Transport { get; set; } = string.Empty;
 
         internal override void Write(XmlWriter writer)
         {
+            writer.WriteStartElement(PrefixConstants.SOAP, "binding", NamespaceConstants.SOAP);
 
+            if (Style != SoapBindingStyle.Default)
+                writer.WriteAttributeString("style", Style == SoapBindingStyle.Rpc ? "rpc" : "document");
+
+            writer.WriteAttributeString("transport", Transport);
+
+            writer.WriteEndElement();
         }
     }
 }